backend/requirements.txt
anyio==3.6.2
fastapi==0.95.1
idna==3.4
pydantic==1.10.7
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
greenlet==2.0.2
psycopg2-binary==2.9.6
sqlalchemy==2.0.13
Mako==1.2.4
MarkupSafe==2.1.2
alembic==1.10.4
click==8.1.3
h11==0.14.0
uvicorn==0.22.0
certifi==2023.5.7
charset-normalizer==3.1.0
requests==2.30.0
urllib3==2.0.2
lxml==4.9.2
beautifulsoup4==4.12.2
bs4==0.0.1
soupsieve==2.4.1
python-dotenv==1.0.0
